tailorcv-mcp/requirements.txt
# LangChain core
langchain
langchain-google-genai
langchain-groq
langchain-community
langgraph

# PDF & DOCX processing
pypdf
python-docx

# Vector DB (local)
chromadb
langchain-chroma

# Utilities
python-dotenv
tiktoken
pydantic

# REST API (tailorcv-ui)
fastapi
uvicorn[standard]